Controllers return types

See merge request cse2115/2020-2010/7-student-house-food-management/op26-sem51!32

+++ requests/src/main/java/nl/tudelft/sem/requests/controllers/HouseController.java
@@ -1,8 +1,10 @@
 package nl.tudelft.sem.requests.controllers;
 
+import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 import nl.tudelft.sem.requests.config.Username;
 import nl.tudelft.sem.requests.entities.House;
 import nl.tudelft.sem.requests.entities.User;
@@ -66,10 +68,10 @@ public class HouseController {
      */
     @GetMapping("/getHouse/{houseNumber}")
     @ResponseBody
-    public Optional<House> getHouseByHouseNumber(@PathVariable int houseNumber,
-                                                 @Username String username) {
-        System.out.println(username);
-        return houseRepository.findById(houseNumber);
+    public ResponseEntity<House> getHouseByHouseNumber(@PathVariable int houseNumber) {
+        Optional<House> house = houseRepository.findById(houseNumber);
+        return house.map(value -> ResponseEntity.ok().body(value))
+                .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     /**
@@ -79,13 +81,10 @@ public class HouseController {
      * @return a list of users in this house
      */
     @GetMapping("/getUsersFromHouse/{houseNumber}")
-    public List<User> getAllUsersFromHouse(@PathVariable int houseNumber) {
+    public ResponseEntity<List<User>> getAllUsersFromHouse(@PathVariable int houseNumber) {
         Optional<House> house = houseRepository.findById(houseNumber);
-        List<User> users = new ArrayList<>();
-        if (house.isPresent()) {
-            users.addAll(house.get().getUsers());
-        }
-        return users;
+        return house.map(value -> ResponseEntity.ok().body(List.copyOf(value.getUsers())))
+                .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     /**
@@ -94,14 +93,16 @@ public class HouseController {
      * @param house    house to be added
      * @param username username of the user creating the house
      */
-    @PostMapping(value = "/addNewHouse", consumes = "application/json")
-    //@RequestMapping(value = "/addNewHouse", method=RequestMethod.POST,
-    //       consumes=MediaType.APPLICATION_JSON_VALUE, produces=MediaType.APPLICATION_JSON_VALUE)
-    public void addNewHouse(@RequestBody House house, @RequestParam("username") String username) {
-        houseRepository.save(house);
+    @PostMapping("/addNewHouse")
+    public ResponseEntity<String> addNewHouse(@RequestBody House house, @Username String username) {
         Optional<User> user = userRepository.findById(username);
-        user.ifPresent(u -> u.setHouse(house));
-        user.ifPresent(u -> userRepository.save(u));
+        return user.map(u -> {
+            houseRepository.save(house);
+            u.setHouse(house);
+            userRepository.save(u);
+            return ResponseEntity.created(URI.create("/addNewHouse")).body("House was created");
+        }).orElseGet(() ->
+                ResponseEntity.badRequest().body("User is not present in the database."));
     }
 
     /**
@@ -113,21 +114,22 @@ public class HouseController {
      *         INTERNAL_SERVER_ERROR - the house couldn't be updated because of a server error
      */
     @PutMapping("/updateHouse")
-    public ResponseEntity<House> updateHouse(@RequestBody House houseWithNewInfo) {
+    public ResponseEntity<String> updateHouse(@RequestBody House houseWithNewInfo) {
         Optional<House> house = houseRepository.findById(houseWithNewInfo.getHouseNr());
 
         if (house.isPresent()) {
             try {
+                // update the house
                 houseRepository.save(houseWithNewInfo);
             } catch (Exception e) {
-                return new ResponseEntity("House couldn't be updated!",
+                return new ResponseEntity<>("House couldn't be updated!",
                     HttpStatus.INTERNAL_SERVER_ERROR);
             }
 
-            return new ResponseEntity("House updated successfully!", HttpStatus.OK);
+            return new ResponseEntity<>("House updated successfully!", HttpStatus.OK);
         }
 
-        return new ResponseEntity("House not found!", HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>("House not found!", HttpStatus.NOT_FOUND);
     }
 
     /**
@@ -136,21 +138,25 @@ public class HouseController {
      * @param houseNumber houseNumber of the house to delete from the database
      */
     @DeleteMapping("/deleteHouse/{houseNumber}")
-    public void deleteHouse(@PathVariable int houseNumber) {
+    public ResponseEntity<String> deleteHouse(@PathVariable int houseNumber) {
         Optional<House> house = houseRepository.findById(houseNumber);
         if (house.isPresent()) {
-            List<User> users = getAllUsersFromHouse(houseNumber);
+            ResponseEntity<List<User>> usersResponse = getAllUsersFromHouse(houseNumber);
+            if (usersResponse.getStatusCode() != HttpStatus.OK) {
+                return ResponseEntity.badRequest().body("Invalid house number.");
+            }
+            List<User> users = usersResponse.getBody();
             for (User user : users) {
                 user.setHouse(null);
                 userRepository.save(user);
             }
             houseRepository.deleteById(houseNumber);
-            System.out.println("house successfully deleted");
+            return ResponseEntity.ok().body("House successfully deleted.");
         } else {
-            System.out.println("house not found!");
+            return ResponseEntity.notFound().build();
         }
     }
-
+    
     /**
      * User joining a house, once a request has been approved.
      *
@@ -163,15 +169,94 @@ public class HouseController {
 
         // set the new house of the user
         user.get().setHouse(house.get());
-
+        
         UserController userController = new UserController(userRepository);
-        userController.updateUser(user.get());
+        userController.updateUser(user.get(), user.get().getUsername());
+        // userController.updateUser(user.get());
 
-        // add the user to the members
-        house.get().getUsers().add(user.get());
+        // // add the user to the members
+        // house.get().getUsers().add(user.get());
 
-        HouseController houseController = new HouseController(houseRepository, userRepository);
-        houseController.updateHouse(house.get());
+        // HouseController houseController = new HouseController(houseRepository, userRepository);
+        // houseController.updateHouse(house.get());
+        
+    }
+    
+    /**
+     * Method for subtracting credits, when products is expired.
+     *
+     * @param username username of the user that has reported the product as expired
+     * @param credits  amount of credits to be split
+     * @return true if the credits were subtracted from each user
+     */
+    @PostMapping("/splitCreditsExpired")
+    public @ResponseBody
+    ResponseEntity<?> splitCreditsWhenExpired(@RequestParam String username,
+                                              @RequestParam float credits) {
+        
+        // @ResponseBody means the returned String is the response, not a view name
+        // @RequestParam means it is a parameter from the GET or POST request
+        
+        
+        User currentUser = userRepository.findByUsername(username);
+        
+        House house = currentUser.getHouse();
+        
+        Set<User> users = house.getUsers();
+        
+        if (users == null || users.isEmpty()) {
+            return ResponseEntity.badRequest().build();
+        }
+        
+        credits = credits / users.size();
+        
+        try {
+            for (User user : users) {
+                if (userRepository.updateUserCredits(user.getHouse().getHouseNr(),
+                        user.getEmail(),
+                        user.getTotalCredits() - credits,
+                        user.getUsername()) == 0) { //NOPMD
+                    return ResponseEntity.badRequest().build();
+                }
+                
+            }
+            return ResponseEntity.created(URI.create("/editUserCredits")).build();
+            
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+
+    /**Get all products from a fridge of the house.
+     *
+     *@param houseNr the number of the house which products from the fridge will be displayed
+     *@return all products from the fridge
+     */
+    @PostMapping("/getUsernamesByHouse")
+    public @ResponseBody
+    ResponseEntity<?> getUsernamesByHouse(@RequestParam int houseNr) {
+        
+        // @ResponseBody means the returned String is the response, not a view name
+        // @RequestParam means it is a parameter from the GET or POST request
+        
+        House house = houseRepository.findByHouseNr(houseNr);
+        Set<User> users = house.getUsers();
+        
+        
+        if (users == null || users.isEmpty() || house == null) {
+            return ResponseEntity.badRequest().build();
+        }
+        
+        List<String> usernames = new ArrayList<>();
+        for (User user : users) {
+            usernames.add(user.getUsername());
+        }
+        
+        try {
+            return ResponseEntity.created(URI.create("/getUsernamesByHouse")).body(usernames);
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
 
     }
 
@@ -184,9 +269,9 @@ public class HouseController {
      *         FORBIDDEN - the house number of the user is different from the one given
      *         NOT_FOUND - if the user or the house do not exist in the database
      */
-    @PutMapping("/leaveHouse/{username}/{houseNumber}")
-    public ResponseEntity<House> userLeavingHouse(@PathVariable String username,
-                                                  @PathVariable int houseNumber) {
+    @PutMapping("/leaveHouse/{houseNumber}")
+    public ResponseEntity<String> userLeavingHouse(@Username String username,
+                                                   @PathVariable int houseNumber) {
         Optional<House> house = houseRepository.findById(houseNumber);
         Optional<User> user = userRepository.findById(username);
 
@@ -201,19 +286,20 @@ public class HouseController {
                         deleteHouse(houseNumber);
                     }
 
-                    return new ResponseEntity("You successfully removed " + username
+                    return new ResponseEntity<>("You successfully removed " + username
                         + " from house number " + houseNumber + "!", HttpStatus.OK);
                 }
 
-                return new ResponseEntity("You can not remove a user from a different household!",
+                return new ResponseEntity<>("You can not remove a user from a different household!",
                     HttpStatus.FORBIDDEN);
             }
 
-            return new ResponseEntity("The user does not have a house!",
+            return new ResponseEntity<>("The user does not have a house!",
                 HttpStatus.FORBIDDEN);
         }
 
-        return new ResponseEntity("The user or the house were not found, please check again!",
+        return new ResponseEntity<>("The user or the house were not found, please check again!",
             HttpStatus.NOT_FOUND);
     }
 }
+
